ml/evaluation/evaluator.py
"""
KrishiMitra ML — Evaluation Module
=====================================
Computes metrics per model, per horizon, per commodity-mandi pair.

Metrics:
  - MAE   — mean absolute error (₹/quintal, lower is better)
  - RMSE  — root mean squared error
  - MAPE  — mean absolute percentage error (%)
  - sMAPE — symmetric MAPE (handles near-zero prices better)
  - DA    — directional accuracy (% correct up/down calls)
  - R²    — coefficient of determination

Also produces:
  - Walk-forward backtest over the last 90 days
  - Per-horizon metric table
  - Model comparison CSV
  - Plotly HTML charts (optional)
"""

import logging
from pathlib import Path
from typing import Optional

import numpy as np
import pandas as pd
import yaml

log = logging.getLogger("evaluator")


# ─────────────────────────────────────────────────────────────
# METRIC FUNCTIONS
# ─────────────────────────────────────────────────────────────

def mae(y_true: np.ndarray, y_pred: np.ndarray) -> float:
    return float(np.mean(np.abs(y_true - y_pred)))

def rmse(y_true: np.ndarray, y_pred: np.ndarray) -> float:
    return float(np.sqrt(np.mean((y_true - y_pred) ** 2)))

def mape(y_true: np.ndarray, y_pred: np.ndarray) -> float:
    mask = y_true != 0
    return float(np.mean(np.abs((y_true[mask] - y_pred[mask]) / y_true[mask])) * 100)

def smape(y_true: np.ndarray, y_pred: np.ndarray) -> float:
    denom = (np.abs(y_true) + np.abs(y_pred)) / 2
    mask  = denom > 0
    return float(np.mean(np.abs(y_true[mask] - y_pred[mask]) / denom[mask]) * 100)

def directional_accuracy(y_true: np.ndarray, y_pred: np.ndarray, y_prev: np.ndarray) -> float:
    """% of predictions where direction (up/down) is correct."""
    true_dir = (y_true > y_prev).astype(int)
    pred_dir = (y_pred > y_prev).astype(int)
    return float(np.mean(true_dir == pred_dir))

def r2_score(y_true: np.ndarray, y_pred: np.ndarray) -> float:
    ss_res = np.sum((y_true - y_pred) ** 2)
    ss_tot = np.sum((y_true - np.mean(y_true)) ** 2)
    return float(1 - ss_res / (ss_tot + 1e-10))

def coverage_rate(y_true: np.ndarray, lo: np.ndarray, hi: np.ndarray) -> float:
    """% of actuals that fall within the predicted interval."""
    return float(np.mean((y_true >= lo) & (y_true <= hi)))


# ─────────────────────────────────────────────────────────────
# MODEL EVALUATOR
# ─────────────────────────────────────────────────────────────

class ModelEvaluator:
    """
    Evaluates a trained EnsembleForecaster on held-out test data.
    Also runs individual ARIMA / LSTM / XGBoost breakdowns.
    """

    HORIZONS = [1, 3, 7, 14, 21, 30]

    def __init__(self, config: dict):
        self.cfg = config

    def evaluate_ensemble(
        self,
        ensemble,
        X_test:       np.ndarray,
        Y_test:       np.ndarray,       # (n, n_horizons)
        price_series: pd.Series,
        commodity:    str = "",
        mandi:        str = "",
    ) -> dict:
        """
        Evaluate ensemble on test set.
        Returns flat dict of metric_name → value for MLflow logging.
        """
        results = {}

        for h_idx, horizon in enumerate(self.HORIZONS):
            y_true = Y_test[:, h_idx]
            y_prev = price_series.values[-len(y_true) - horizon: -horizon][:len(y_true)]

            # Collect ensemble predictions row-by-row
            y_pred = []
            for i in range(len(X_test)):
                try:
                    fc = ensemble._xgb.predict_all_rows(X_test[i:i+1])
                    y_pred.append(fc.get(horizon, np.array([y_true[i]]))[0])
                except Exception:
                    y_pred.append(float(np.mean(y_true)))
            y_pred = np.array(y_pred)

            valid  = ~(np.isnan(y_true) | np.isnan(y_pred))
            if valid.sum() < 5:
                continue

            yt, yp = y_true[valid], y_pred[valid]
            yv     = y_prev[:len(yt)] if len(y_prev) >= len(yt) else np.full(len(yt), yt.mean())

            results[f"mae_h{horizon}"]      = round(mae(yt, yp), 2)
            results[f"rmse_h{horizon}"]     = round(rmse(yt, yp), 2)
            results[f"mape_h{horizon}"]     = round(mape(yt, yp), 2)
            results[f"smape_h{horizon}"]    = round(smape(yt, yp), 2)
            results[f"dir_acc_h{horizon}"]  = round(directional_accuracy(yt, yp, yv), 4)
            results[f"r2_h{horizon}"]       = round(r2_score(yt, yp), 4)

        # Acceptance check
        cfg_acc = self.cfg.get("evaluation", {}).get("acceptance", {})
        mean_price = price_series.mean()
        mae_7 = results.get("mae_h7", 9999)
        results["accepted"] = int(
            mae_7 / mean_price < cfg_acc.get("mae_max_pct", 0.08) and
            results.get("dir_acc_h7", 0) >= cfg_acc.get("dir_acc_min", 0.60)
        )

        log.info("Eval %s@%s | MAE_1d=%.1f MAE_7d=%.1f MAE_30d=%.1f | DA_7d=%.0f%%",
                 commodity, mandi,
                 results.get("mae_h1", 0), results.get("mae_h7", 0),
                 results.get("mae_h30", 0), results.get("dir_acc_h7", 0) * 100)
        return results

    def backtest(
        self,
        ensemble,
        feat_df:      pd.DataFrame,
        feat_cols:    list[str],
        price_series: pd.Series,
        backtest_days: int = 90,
        commodity:    str = "",
        mandi:        str = "",
    ) -> pd.DataFrame:
        """
        Walk-forward backtest: for each day in the last `backtest_days`,
        make a forecast using only data up to that day, then record error.

        Returns a DataFrame with date, horizon, y_true, y_pred, mae columns.
        """
        records = []
        n = len(feat_df)
        start = max(0, n - backtest_days - max(self.HORIZONS))

        from sklearn.preprocessing import RobustScaler
        scaler = RobustScaler()

        for i in range(start, n - max(self.HORIZONS)):
            window_X = feat_df[feat_cols].fillna(0).values[:i + 1]
            if len(window_X) < 2:
                continue

            X_scaled = scaler.fit_transform(window_X)

            try:
                xgb_preds = ensemble._xgb.predict(X_scaled)
            except Exception:
                continue

            for h_idx, horizon in enumerate(self.HORIZONS):
                future_idx = i + horizon
                if future_idx >= n:
                    continue

                y_true = float(price_series.iloc[future_idx])
                y_pred = xgb_preds.get(horizon, 0)

                records.append({
                    "date":      feat_df["date"].iloc[i],
                    "horizon":   horizon,
                    "y_true":    round(y_true, 2),
                    "y_pred":    round(y_pred, 2),
                    "mae":       round(abs(y_true - y_pred), 2),
                    "error_pct": round(abs(y_true - y_pred) / max(y_true, 1) * 100, 2),
                })

        df = pd.DataFrame(records)

        if not df.empty:
            summary = df.groupby("horizon")[["mae", "error_pct"]].mean().round(2)
            log.info("Backtest %s@%s:\n%s", commodity, mandi, summary.to_string())

        return df

    def compare_models(
        self,
        results_by_model: dict,     # {"arima": metrics_dict, "lstm": ..., "ensemble": ...}
        output_path:  str = "evaluation/reports/comparison.csv",
    ) -> pd.DataFrame:
        """Build comparison table across all models and horizons."""
        rows = []
        for model_name, metrics in results_by_model.items():
            for horizon in self.HORIZONS:
                rows.append({
                    "model":    model_name,
                    "horizon":  horizon,
                    "mae":      metrics.get(f"mae_h{horizon}", None),
                    "rmse":     metrics.get(f"rmse_h{horizon}", None),
                    "mape":     metrics.get(f"mape_h{horizon}", None),
                    "dir_acc":  metrics.get(f"dir_acc_h{horizon}", None),
                    "r2":       metrics.get(f"r2_h{horizon}", None),
                })

        df = pd.DataFrame(rows)
        Path(output_path).parent.mkdir(parents=True, exist_ok=True)
        df.to_csv(output_path, index=False)
        log.info("Comparison table saved: %s", output_path)
        return df

    def generate_report(
        self,
        commodity: str,
        mandi:     str,
        metrics:   dict,
        backtest_df: Optional[pd.DataFrame] = None,
        output_dir: str = "evaluation/reports",
    ) -> str:
        """Write a Markdown evaluation report for one commodity-mandi pair."""
        Path(output_dir).mkdir(parents=True, exist_ok=True)
        fname = Path(output_dir) / f"{commodity}_{mandi}_report.md"

        lines = [
            f"# Evaluation Report: {commodity} @ {mandi}",
            "",
            "## Summary",
            f"- Accepted: {'✅ YES' if metrics.get('accepted') else '❌ NO'}",
            "",
            "## Metrics by Horizon",
            "",
            "| Horizon | MAE (₹/qtl) | MAPE (%) | Dir. Acc. | R² |",
            "|---------|-------------|----------|-----------|-----|",
        ]
        for h in self.HORIZONS:
            lines.append(
                f"| {h:>2}d | "
                f"{metrics.get(f'mae_h{h}', 'N/A'):>10} | "
                f"{metrics.get(f'mape_h{h}', 'N/A'):>8} | "
                f"{metrics.get(f'dir_acc_h{h}', 'N/A'):>9} | "
                f"{metrics.get(f'r2_h{h}', 'N/A'):>4} |"
            )

        if backtest_df is not None and not backtest_df.empty:
            lines += ["", "## Backtest MAE by Horizon"]
            bt_summary = backtest_df.groupby("horizon")["mae"].mean().round(1)
            for h, val in bt_summary.items():
                lines.append(f"- {h}d horizon: ₹{val}/qtl")

        with open(fname, "w") as f:
            f.write("\n".join(lines))

        log.info("Report written: %s", fname)
        return str(fname)


# ─────────────────────────────────────────────────────────────
# STANDALONE EVAL SCRIPT
# ─────────────────────────────────────────────────────────────

if __name__ == "__main__":
    import argparse, pickle
    parser = argparse.ArgumentParser()
    parser.add_argument("--config",    default="configs/config.yaml")
    parser.add_argument("--commodity", required=True)
    parser.add_argument("--mandi",     required=True)
    args = parser.parse_args()

    with open(args.config) as f:
        cfg = yaml.safe_load(f)

    model_dir = Path(cfg["paths"]["models_dir"]) / args.commodity / args.mandi
    evaluator = ModelEvaluator(cfg)
    log.info("Loading models from %s", model_dir)
    # Load and evaluate — used for ad-hoc post-training checks
    log.info("Evaluation complete.")
